Feat: "Completed Hyperparameter Optimization Solution"
This commit includes the solution to the exercise from the 'Python: Working with Predictive Analytics' course on LinkedIn Learning, specifically the section on hyperparameter optimization. This exercise provided an opportunity to apply the skills learned in the course, particularly in optimizing hyperparameters for machine learning models. The experience gained from this exercise will be beneficial in achieving optimal model performance and improving predictive analytics outcomes in future data science projects.